Moved GetProductsById functionality to separate handler and endpoint
