+++ src/catalog/Catalog.API/Features/GetProductById/GetProductByIdEndpoint.cs
@@ -1,0 +1,23 @@
+﻿namespace Catalog.API.Features.GetProductById
+{
+    public class GetProductByIdEndpoint : ICarterModule
+    {
+        public record GetProductResponse(Product Product);
+
+        public void AddRoutes(IEndpointRouteBuilder app)
+        {
+            app.MapGet("/products/{id}", async (Guid id, ISender sender) =>
+            {
+                var result = await sender.Send(new GetProductByIdQuery(id));
+
+                var response = result.Adapt<GetProductResponse>();
+                return Results.Ok(response);
+            })
+            .WithName("GetProduct")
+            .Produces<GetProductResponse>(StatusCodes.Status200OK)
+            .ProducesProblem(StatusCodes.Status400BadRequest)
+            .WithSummary("Get Product")
+            .WithDescription("Get Product By Id");
+        }
+    }
+}
